requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
Django==3.2.18
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-env==0.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
httpie==3.2.1
idna==3.4
importlib-metadata==6.0.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.4.0
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7.1
pyxios==1.0.2
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
sqlparse==0.4.3
typing-extensions==4.5.0
urllib3==1.26.14
whitenoise==6.4.0
zipp==3.15.0
